_project/requirement.txt

aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioredis==2.0.1
aiosignal==1.4.0
amazoncaptcha==0.5.11
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.5
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
kombu==5.5.4
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.3.8
playwright==1.54.0
prompt_toolkit==3.0.51
propcache==0.3.2
pyee==13.0.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.30.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1